test(wiki): split status report tests under max-lines limit
The 363-line test file exceeded the 300-line max-lines rule. Extract
shared fixture helpers to tests/helpers/__fixtures__/ and split the
verdict-derivation tests into a sibling file. All 6 tests still pass.
